babel: extract stylesheet
